requirements.txt
playwright>=1.43.0
opencv-python>=4.9.0
easyocr>=1.7.1
pyautogui>=0.9.54
numpy>=1.26.0
Pillow>=10.3.0
requests>=2.31.0
keyboard>=0.13.5
